Handle case where resources do not exist gracefully

+++ migrations/20240530082301_publish_users_to_nats.ts
@@ -48,9 +48,13 @@ export async function up(knex: Knex): Promise<void> {
 }
 
 export async function down(knex: Knex): Promise<void> {
-  await knex.schema.table("users", table => {
-    table.dropColumn("last_seq");
-  });
+  if (await knex.schema.hasColumn("users", "last_seq")) {
+    await knex.schema.table("users", table => {
+      table.dropColumn("last_seq");
+    });
+  }
 
-  await knex.schema.dropTable("user_ids");
+  if (await knex.schema.hasTable("user_ids")) {
+    await knex.schema.dropTable("user_ids");
+  }
 }
